+++ src/equations/numerical_fluxes.jl
@@ -154,18 +154,18 @@ function Base.show(io::IO, d::DissipationGlobalLaxFriedrichs)
 end
 
 """
-    DissipationLocalLaxFriedrichs(max_abs_speed=max_abs_speed_naive)
+    DissipationLocalLaxFriedrichs(max_abs_speed=max_abs_speed)
 
 Create a local Lax-Friedrichs dissipation operator where the maximum absolute wave speed
 is estimated as
 `max_abs_speed(u_ll, u_rr, orientation_or_normal_direction, equations)`,
-defaulting to [`max_abs_speed_naive`](@ref).
+defaulting to [`max_abs_speed`](@ref).
 """
 struct DissipationLocalLaxFriedrichs{MaxAbsSpeed}
     max_abs_speed::MaxAbsSpeed
 end
 
-DissipationLocalLaxFriedrichs() = DissipationLocalLaxFriedrichs(max_abs_speed_naive)
+DissipationLocalLaxFriedrichs() = DissipationLocalLaxFriedrichs(max_abs_speed)
 
 @inline function (dissipation::DissipationLocalLaxFriedrichs)(u_ll, u_rr,
                                                               orientation_or_normal_direction,
@@ -188,6 +188,8 @@ Simple and fast estimate of the maximal wave speed of the Riemann problem with l
 
 For non-integer arguments `normal_direction` in one dimension, `max_abs_speed_naive` returns
 `abs(normal_direction[1]) * max_abs_speed_naive(u_ll, u_rr, 1, equations)`.
+
+Slightly more diffusive/overestimating than [`max_abs_speed`](@ref).
 """
 function max_abs_speed_naive end
 
@@ -211,6 +213,8 @@ i.e., the wave speeds used in `max_dt` which computes the maximum stable time st
 
 For non-integer arguments `normal_direction` in one dimension, `max_abs_speed_naive` returns
 `abs(normal_direction[1]) * max_abs_speed_naive(u_ll, u_rr, 1, equations)`.
+
+Defaults to [`min_max_speed_naive`](@ref) if no specialized version for the 'equations` at hand is available.
 """
 @inline function max_abs_speed(u_ll, u_rr,
                                orientation_or_normal_direction,
@@ -222,13 +226,13 @@ end
 const FluxLaxFriedrichs{MaxAbsSpeed} = FluxPlusDissipation{typeof(flux_central),
                                                            DissipationLocalLaxFriedrichs{MaxAbsSpeed}}
 """
-    FluxLaxFriedrichs(max_abs_speed=max_abs_speed_naive)
+    FluxLaxFriedrichs(max_abs_speed=max_abs_speed)
 
 Local Lax-Friedrichs (Rusanov) flux with maximum wave speed estimate provided by
 `max_abs_speed`, cf. [`DissipationLocalLaxFriedrichs`](@ref) and
 [`max_abs_speed_naive`](@ref).
 """
-function FluxLaxFriedrichs(max_abs_speed = max_abs_speed_naive)
+function FluxLaxFriedrichs(max_abs_speed = max_abs_speed)
     FluxPlusDissipation(flux_central, DissipationLocalLaxFriedrichs(max_abs_speed))
 end
 
@@ -244,7 +248,7 @@ See [`FluxLaxFriedrichs`](@ref).
 const flux_lax_friedrichs = FluxLaxFriedrichs()
 
 @doc raw"""
-    DissipationLaxFriedrichsEntropyVariables(max_abs_speed=max_abs_speed_naive)
+    DissipationLaxFriedrichsEntropyVariables(max_abs_speed=max_abs_speed)
 
 Create a local Lax-Friedrichs-type dissipation operator that is provably entropy stable. This operator
 must be used together with an entropy-conservative two-point flux function (e.g., `flux_ec`) to yield 
@@ -259,7 +263,7 @@ f^{\mathrm{ES}} = f^{\mathrm{EC}} - \frac{1}{2} \lambda_{\mathrm{max}} H (w_r - 
 ```
 where ``f^{\mathrm{EC}}`` is the entropy-conservative two-point flux function (computed with, e.g., `flux_ec`), ``\lambda_{\mathrm{max}}`` 
 is the maximum wave speed estimated as `max_abs_speed(u_l, u_r, orientation_or_normal_direction, equations)`,
-defaulting to [`max_abs_speed_naive`](@ref), ``H`` is a symmetric positive-definite dissipation matrix that
+defaulting to [`max_abs_speed`](@ref), ``H`` is a symmetric positive-definite dissipation matrix that
 depends on the left and right states `u_l` and `u_r`, and ``(w_r - w_l)`` is the jump in entropy variables.
 Ideally, ``H (w_r - w_l) = (u_r - u_l)``, such that the dissipation operator is consistent with the local
 Lax-Friedrichs dissipation.
@@ -277,7 +281,7 @@ struct DissipationLaxFriedrichsEntropyVariables{MaxAbsSpeed}
     max_abs_speed::MaxAbsSpeed
 end
 
-DissipationLaxFriedrichsEntropyVariables() = DissipationLaxFriedrichsEntropyVariables(max_abs_speed_naive)
+DissipationLaxFriedrichsEntropyVariables() = DissipationLaxFriedrichsEntropyVariables(max_abs_speed)
 
 function Base.show(io::IO, d::DissipationLaxFriedrichsEntropyVariables)
     print(io, "DissipationLaxFriedrichsEntropyVariables(", d.max_abs_speed, ")")